adjust imports after src/utils.js moved to src/utils/index.js
