+++ src/components/VmDetails/cards/UtilizationCard/DiskCharts.js
@@ -13,8 +13,7 @@ import {
   DonutChart,
 } from 'patternfly-react'
 
-import { convertValueMap } from '../../../../utils/storage-conversion'
-import { round } from '../../../../utils/round'
+import { convertValueMap, round } from '../../../../utils'
 
 import style from './style.css'
